HOCS-2419: correct private office check
Currently when doing a template download we try and receive a 
Private Office team no matter what the case type is. This 
call does not work for case types other than 'MIN', and 'DTEN' as 
they do not contain a Private Office team for the topic. This 
change makes sure that we only send a call for the topic 
team when we are using cases that have a Private Office. Tests 
have been added to accompany this test to make sure we can 
create the file.