HOCS-2419: unable to download TRO template (#21)
* HOCS-2419: add ds_store to gitignore

Currently .DS_Store files are not excluded from git. This change 
enforces that these are excluded from git so they cannot be 
accidentally staged.

* HOCS-2419: correct private office check

Currently when doing a template download we try and receive a 
Private Office team no matter what the case type is. This 
call does not work for case types other than 'MIN', and 'DTEN' as 
they do not contain a Private Office team for the topic. This 
change makes sure that we only send a call for the topic 
team when we are using cases that have a Private Office. Tests 
have been added to accompany this test to make sure we can 
create the file.

* HOCS-2419: change to use the new info endpoint

A number endpoint has been exposed by info-service 
to assist with retrieving a case types stages. 
This is then used within templates to check whether there 
is a private office stage for the case type requested.

* HOCS-2419: change postcode to a not real one

Co-authored-by: Douglas Gardner <douglas.gardner@digital.homeoffice.gov.uk>

Co-authored-by: Douglas Gardner <douglas.gardner@digital.homeoffice.gov.uk>