HOCS-2576: refactor drone script
As we don't need to push branch name unless we are doing 
a epic or hotfix branch this has been pulled out to a 
different step. Reorder deploy so that the clone of the kube 
repo is successful before we tag the release. Only clone the 
kube project if we are doing a promotion. Move the reading 
of the version to the docker-semver-tag step as it is part of 
that. Indent the commands to be more readable.