HOCS-2576: fix deploy environments
At present QA would not deploy correctly as it depends on a 
DRONE_DEPLOY_TO environment variable. As we are 
releasing to both cs-qa and wcs-qa simultaneously we don't 
need this. We currently have 2 steps that handle prod deployment, 
this commit also simplifies this by combining both. As 
environment variables are not passed between pipeline steps I 
have reverted the change to put this in the semver step.