HOCS-2576: refactor code further
To make the drone UI more readable this commit changes 
the name of the steps. This has had a knock on impact 
where we need to change the depends_on to accommodate  
this. Further changes include, pushing a branch to Quay on 
every push, removing unnecessary secrets from steps, and 
add depends_on for build steps to allow for parallelism.