HOCS-2576 Remove explicit trigger definitions
We're being slightly too clever with the step definitions, and
accidentally overwrote the "trigger this on promote" with the "trigger
this on push to master" condition, meaning that on a promote Drone
didn't know what to do.

This commit removes the trigger stuff from the deploy step, because it
is introducing complexity that we don't fully understand for only a
minor benefit.