HOCS-2576: generalise prod deployment (#31)
* HOCS-2576: generalise prod deployment

At present we only allow for deployment to production environments that
are either `cs-prod` or `wcs-prod`. This change allows to release to any prod environment specified.

* HOCS-2576: change secret name

At present we use `QUAY_ROBOT_PASSWORD` for the quay secret. This
is not strictly what this is, so to make it clear the secrets name has been
changed to `QUAY_ROBOT_TOKEN`.