HOCS-2657: change escaping implementation
Currently we use a HTML escaping function for the variables, 
however as we use a XML library this fails. This is because é and 
some other characters are not valid XML when escaped using HTML 
these fail the replace. This change uses the Apache Commons  
`StringEscapeUtils.escapeXml` functionality to escape to the entity 
number instead of the entity code.