HOCS-1746 Use ACP Quay.io proxy (#7)
This commit changes references to Quay.io to a Home Office hosted proxy
of Quay.io. This will mean that if Quay.io is unavailable, previously
pulled images will still be available for download. However, it would
still not be possible to push new images, which must go to Quay
directly.

This merely moves the single point of failure, but to one that is within
the Home Office's responsibility and therefore might be easier to fix.

If the ACP mirror goes down, you can revert this commit without
consequence.

Co-authored-by: Nick McMahon <nick.mcmahon@digital.homeoffice.gov.uk>